ftx,ftxus: convert to TX
