--- src/exchanges/ftx-client.js
@@ -1,9 +1,0 @@
-const FtxBaseClient = require("./ftx-base");
-
-class FtxClient extends FtxBaseClient {
-  constructor({ wssPath = "wss://ftx.com/ws", watcherMs } = {}) {
-    super({ name: "FTX", wssPath, watcherMs });
-  }
-}
-
-module.exports = FtxClient;
